Moving SimpleTestPackage helpers to test utility
