+++ src/NuGet.Core/NuGet.Test.Utility/project.json
@@ -4,7 +4,8 @@
     "warningsAsErrors": true
   },
   "dependencies": {
-    "xunit": "2.1.0"
+    "xunit": "2.1.0",
+    "NuGet.Packaging": "3.4.0-*"
   },
   "frameworks": {
     "net45": {
