+++ test/NuGet.Core.Tests/NuGet.Commands.Test/IncludeTypeTests.cs
@@ -1,17 +1,13 @@
 ﻿using System.Collections.Generic;
 using System.IO;
-using System.IO.Compression;
 using System.Linq;
-using System.Text;
 using System.Threading.Tasks;
 using System.Xml.Linq;
 using NuGet.Configuration;
 using NuGet.Frameworks;
 using NuGet.LibraryModel;
-using NuGet.Packaging.Core;
 using NuGet.ProjectModel;
 using NuGet.Test.Utility;
-using NuGet.Versioning;
 using Xunit;
 
 namespace NuGet.Commands.Test
@@ -460,14 +456,14 @@ namespace NuGet.Commands.Test
                         }
                  }";
 
-                var x = new TestPackage()
+                var x = new SimpleTestPackageContext()
                 {
                     Id = "packageX",
                     Version = "1.0.0",
                     Include = "runtime"
                 };
 
-                var z = new TestPackage()
+                var z = new SimpleTestPackageContext()
                 {
                     Id = "packageZ",
                     Version = "1.0.0",
@@ -476,7 +472,7 @@ namespace NuGet.Commands.Test
 
                 z.Dependencies.Add(x);
 
-                var y = new TestPackage()
+                var y = new SimpleTestPackageContext()
                 {
                     Id = "packageY",
                     Version = "1.0.0",
@@ -487,12 +483,12 @@ namespace NuGet.Commands.Test
 
                 x.Dependencies.Add(y);
 
-                var packages = new List<TestPackage>()
+                var packages = new List<SimpleTestPackageContext>()
                 {
                     x
                 };
 
-                CreatePackages(packages, Path.Combine(workingDir, "repository"));
+                SimpleTestPackageUtility.CreatePackages(packages, Path.Combine(workingDir, "repository"));
 
                 // Act
                 var result = await StandardSetup(workingDir, logger, projectJson);
@@ -534,14 +530,14 @@ namespace NuGet.Commands.Test
                         }
                  }";
 
-                var z1 = new TestPackage()
+                var z1 = new SimpleTestPackageContext()
                 {
                     Id = "packageZ",
                     Version = "1.0.0",
                     Include = "runtime"
                 };
 
-                var y1 = new TestPackage()
+                var y1 = new SimpleTestPackageContext()
                 {
                     Id = "packageY",
                     Version = "1.0.0",
@@ -549,14 +545,14 @@ namespace NuGet.Commands.Test
                 };
                 y1.Dependencies.Add(z1);
 
-                var z2 = new TestPackage()
+                var z2 = new SimpleTestPackageContext()
                 {
                     Id = "packageZ",
                     Version = "2.0.0",
                     Include = "compile"
                 };
 
-                var y2 = new TestPackage()
+                var y2 = new SimpleTestPackageContext()
                 {
                     Id = "packageY",
                     Version = "2.0.0",
@@ -565,7 +561,7 @@ namespace NuGet.Commands.Test
 
                 y2.Dependencies.Add(z2);
 
-                var a = new TestPackage()
+                var a = new SimpleTestPackageContext()
                 {
                     Id = "packageA",
                     Version = "1.0.0"
@@ -573,7 +569,7 @@ namespace NuGet.Commands.Test
 
                 a.Dependencies.Add(y1);
 
-                var b = new TestPackage()
+                var b = new SimpleTestPackageContext()
                 {
                     Id = "packageB",
                     Version = "2.0.0"
@@ -581,13 +577,13 @@ namespace NuGet.Commands.Test
 
                 b.Dependencies.Add(y2);
 
-                var packages = new List<TestPackage>()
+                var packages = new List<SimpleTestPackageContext>()
                 {
                     a,
                     b
                 };
 
-                CreatePackages(packages, Path.Combine(workingDir, "repository"));
+                SimpleTestPackageUtility.CreatePackages(packages, Path.Combine(workingDir, "repository"));
 
                 // Act
                 var result = await StandardSetup(workingDir, logger, projectJson);
@@ -1516,8 +1512,17 @@ namespace NuGet.Commands.Test
 
             request.ExternalProjects = new List<ExternalProjectReference>()
             {
-                new ExternalProjectReference("TestProject1", spec1, null, new string[] { "TestProject2" }),
-                new ExternalProjectReference("TestProject2", spec2, null, Enumerable.Empty<string>())
+                new ExternalProjectReference(
+                    "TestProject1",
+                    spec1,
+                    Path.Combine(testProject1Dir, "TestProject1.csproj"),
+                    new string[] { "TestProject2" }),
+
+                new ExternalProjectReference(
+                    "TestProject2",
+                    spec2,
+                    Path.Combine(testProject1Dir, "TestProject2.csproj"),
+                    Enumerable.Empty<string>())
             };
 
             var command = new RestoreCommand(logger, request);
@@ -1614,8 +1619,15 @@ namespace NuGet.Commands.Test
             var spec = JsonPackageSpecReader.GetPackageSpec(configJson, "TestProject", specPath);
 
             var request = new RestoreRequest(spec, sources, packagesDir);
-
             request.LockFilePath = Path.Combine(testProjectDir, "project.lock.json");
+
+            request.ExternalProjects.Add(
+                new ExternalProjectReference(
+                    "TestProject",
+                    spec,
+                    Path.Combine(testProjectDir, "TestProject.csproj"),
+                    new string[] { }));
+
 
             var command = new RestoreCommand(logger, request);
 
@@ -1628,25 +1640,25 @@ namespace NuGet.Commands.Test
 
         private static void CreateAToB(string repositoryDir)
         {
-            var b = new TestPackage()
+            var b = new SimpleTestPackageContext()
             {
                 Id = "packageB",
                 Version = "1.0.0"
             };
 
-            var a = new TestPackage()
+            var a = new SimpleTestPackageContext()
             {
                 Id = "packageA",
                 Version = "1.0.0"
             };
             a.Dependencies.Add(b);
 
-            var packages = new List<TestPackage>()
+            var packages = new List<SimpleTestPackageContext>()
             {
                 a
             };
 
-            CreatePackages(packages, repositoryDir);
+            SimpleTestPackageUtility.CreatePackages(packages, repositoryDir);
         }
 
         private static void CreateXYZ(string repositoryDir)
@@ -1656,7 +1668,7 @@ namespace NuGet.Commands.Test
 
         private static void CreateXYZ(string repositoryDir, string include, string exclude)
         {
-            var z = new TestPackage()
+            var z = new SimpleTestPackageContext()
             {
                 Id = "packageZ",
                 Version = "1.0.0",
@@ -1664,7 +1676,7 @@ namespace NuGet.Commands.Test
                 Exclude = exclude
             };
 
-            var y = new TestPackage()
+            var y = new SimpleTestPackageContext()
             {
                 Id = "packageY",
                 Version = "1.0.0",
@@ -1673,7 +1685,7 @@ namespace NuGet.Commands.Test
             };
             y.Dependencies.Add(z);
 
-            var x = new TestPackage()
+            var x = new SimpleTestPackageContext()
             {
                 Id = "packageX",
                 Version = "1.0.0",
@@ -1682,155 +1694,12 @@ namespace NuGet.Commands.Test
             };
             x.Dependencies.Add(y);
 
-            var packages = new List<TestPackage>()
+            var packages = new List<SimpleTestPackageContext>()
             {
                 x
             };
 
-            CreatePackages(packages, repositoryDir);
-        }
-
-        private static FileInfo CreateFullPackage(
-            string repositoryDir,
-            string id,
-            string version)
-        {
-            return CreateFullPackage(repositoryDir, id, version, new List<Packaging.Core.PackageDependency>());
-        }
-
-        private static FileInfo CreateFullPackage(
-            string repositoryDir,
-            string id,
-            string version,
-            Packaging.Core.PackageDependency dependency)
-        {
-            return CreateFullPackage(repositoryDir, id, version,
-                new List<Packaging.Core.PackageDependency>() { dependency });
-        }
-
-        private static FileInfo CreateFullPackage(
-            string repositoryDir,
-            string id,
-            string version,
-            IEnumerable<Packaging.Core.PackageDependency> dependencies)
-        {
-            var file = new FileInfo(Path.Combine(repositoryDir, $"{id}.{version}.nupkg"));
-
-            file.Directory.Create();
-
-            using (var zip = new ZipArchive(File.Create(file.FullName), ZipArchiveMode.Create))
-            {
-                zip.AddEntry("contentFiles/any/any/config.xml", new byte[] { 0 });
-                zip.AddEntry("contentFiles/cs/net45/code.cs", new byte[] { 0 });
-                zip.AddEntry("lib/net45/a.dll", new byte[] { 0 });
-                zip.AddEntry($"build/net45/{id}.targets", @"<targets />", Encoding.UTF8);
-                zip.AddEntry("native/net45/a.dll", new byte[] { 0 });
-                zip.AddEntry("tools/a.exe", new byte[] { 0 });
-
-                var nuspecXml = $@"<?xml version=""1.0"" encoding=""utf-8""?>
-                        <package>
-                        <metadata>
-                            <id>{id}</id>
-                            <version>{version}</version>
-                            <title />
-                            <frameworkAssemblies>
-                                <frameworkAssembly assemblyName=""System.Runtime"" />
-                            </frameworkAssemblies>
-                            <contentFiles>
-                                <files include=""cs/net45/config/config.xml"" buildAction=""none"" />
-                                <files include=""cs/net45/config/config.xml"" copyToOutput=""true"" flatten=""false"" />
-                                <files include=""cs/net45/images/image.jpg"" buildAction=""embeddedresource"" />
-                            </contentFiles>
-                        </metadata>
-                        </package>";
-
-                var xml = XDocument.Parse(nuspecXml);
-
-                if (dependencies.Any())
-                {
-                    var metadata = xml.Element(XName.Get("package")).Element(XName.Get("metadata"));
-
-                    var dependenciesNode = new XElement(XName.Get("dependencies"));
-                    var groupNode = new XElement(XName.Get("group"));
-                    dependenciesNode.Add(groupNode);
-                    metadata.Add(dependenciesNode);
-
-                    foreach (var dependency in dependencies)
-                    {
-                        var node = new XElement(XName.Get("dependency"));
-                        groupNode.Add(node);
-                        node.Add(new XAttribute(XName.Get("id"), dependency.Id));
-                        node.Add(new XAttribute(XName.Get("version"), dependency.VersionRange.ToNormalizedString()));
-
-                        if (dependency.Include.Count > 0)
-                        {
-                            node.Add(new XAttribute(XName.Get("include"), string.Join(",", dependency.Include)));
-                        }
-
-                        if (dependency.Exclude.Count > 0)
-                        {
-                            node.Add(new XAttribute(XName.Get("exclude"), string.Join(",", dependency.Exclude)));
-                        }
-                    }
-                }
-
-                zip.AddEntry($"{id}.nuspec", xml.ToString(), Encoding.UTF8);
-            }
-
-            return file;
-        }
-
-        private class TestPackage
-        {
-            public string Id { get; set; } = "packageA";
-            public string Version { get; set; } = "1.0.0";
-            public List<TestPackage> Dependencies { get; set; } = new List<TestPackage>();
-            public string Include { get; set; } = string.Empty;
-            public string Exclude { get; set; } = string.Empty;
-
-            public PackageIdentity Identity
-            {
-                get
-                {
-                    return new PackageIdentity(Id, NuGetVersion.Parse(Version));
-                }
-            }
-        }
-
-        private static void CreatePackages(List<TestPackage> packages, string repositoryPath)
-        {
-            var done = new HashSet<PackageIdentity>();
-            var toCreate = new Stack<TestPackage>(packages);
-
-            while (toCreate.Count > 0)
-            {
-                var package = toCreate.Pop();
-
-                if (done.Add(package.Identity))
-                {
-                    var dependencies = package.Dependencies.Select(e =>
-                        new Packaging.Core.PackageDependency(
-                            e.Id,
-                            VersionRange.Parse(e.Version),
-                            string.IsNullOrEmpty(e.Include)
-                                ? new List<string>()
-                                : e.Include.Split(',').ToList(),
-                            string.IsNullOrEmpty(e.Exclude)
-                                ? new List<string>()
-                                : e.Exclude.Split(',').ToList()));
-
-                    CreateFullPackage(
-                        repositoryPath,
-                        package.Id,
-                        package.Version,
-                        dependencies);
-
-                    foreach (var dep in package.Dependencies)
-                    {
-                        toCreate.Push(dep);
-                    }
-                }
-            }
+            SimpleTestPackageUtility.CreatePackages(packages, repositoryDir);
         }
 
         private bool IsEmptyFolder(IList<LockFileItem> group)
